Added notebook to do image based data mining in the head and neck.
This notebook has an implementation of IBDM for the HNSCC data looking
at PEG insertion.

It performs a per-voxel t-test only in the region of interest defiled
during the registration, and carries out a permutation test to assess
significance. Visualisation at the end is minimal, and could be
significantly improved!